requirements.txt
pandas==2.2.3
nfl_data_py==0.3.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytest==8.3.4


